feat: frequency response, fixes #9
fixes #9

Compute and plot VCSEL's frequency response H(f) based on the time-domain response y(t) to a small step according to the formula H(f) = FFT(dy(t)/dt). The magnitude of H(f) is computed by multiplying it with its complex conjugate.

A new input current modulation format "small step (3)" was added for that use-case. The time-domain simulation starts at the bias current. The steady-state carrier and photon densities at this point are interpolated from the LI characteristic.

That same technique was applied to the modulation formats "pulse (1)" and "random bits (2)", but not for "step (0)", which starts at I=0.

All changes were implemented in VISTAS1D.py. The files LP_modes.py, VCSEL_params.py and VISTASmodels.py received minor housekeeping updates, and are attached to this commit to provide a working package.