Feat: carrier transport into the QWs, fixes #6, fixes #17#
1. Carrier transport implemented in core algorithm, fixes #6

The core algorithm has been updated with carrier transport through the barriers into the QWs. A light implementation [1] that only adds one equation to the system (no spatial dependency for Nb) was chosen. Carrier transport is now part of the core algorithm, as it is a key mechanism of electrically-pumped VCSELs, that significantly impacts their dynamic response. Still, this mechanism can be "switched off" for, e.g., simulating optically-pumped lasers.

2. Jacobian for both static and dynamic calculations, fixes #17

The Jacobian was updated to account for the new equation describing carriers transport in the barriers Nb. The updated Jacobian is now used for both the static (fsolve) and dynamic (solve_ivp) solutions. It must be noted that solve_ivp actually only makes use of the Jacobian for its stiff-ODE solvers (Radau, BDF and LSODA).

3. New and renamed variables

- N -> Nb, Nw: carrier density in the barriers and in the QWs, respectively. Nb is described with just one additional equation coupled with the equations describing Nw
- ni, nm -> nNb, nNw, nS: number of equations (array dimensions) describing respectively the carrier density in the barrier Nb, the carrier density in the QWs Nw, and the photon densities in each optical mode S
- tauN -> tauNb, tauNw: recombination lifetime in the barriers and QW, respectively
- taub: aggregated non-stimulated recombination lifetime in the barriers: taub = (1/tauNb + 1/tauCap)**-1
- tauw: aggregated non-stimulated recombination lifetime in the barriers: taub = (1/tauNw + 1/tauEsc)**-1
- nqw -> nw: number of quantum wells
- dqw -> dw: quantum well thickness
- Vcav -> Vw: active volume
- Vbar -> Vb: barriers volume
- Vr -> ratio of Vw to Vb
- etaI -> etai: current injection efficiency
- rOx -> rox: oxide aperture
- SCHTransp -> SCHtransp: bool variable describing whether carrier transport should be simulated or not

[1] M. Jungo, D. Erni and W. Baechtold, "Alternative formulation of carrier transport in spatially-dependent laser rate equations", Optical and Quantum Electronics 36(10):881-891, 2004